MAINT: Move grid refinement functions
