+++ src/porepy/grids/refinement.py
@@ -7,8 +7,9 @@ class to generate sets of refined grids.
 from __future__ import annotations
 
 import abc
+import logging
 from pathlib import Path
-from typing import Optional
+from typing import Literal, Optional
 
 import gmsh
 import numpy as np
@@ -20,6 +21,8 @@ from porepy.fracs.fracture_network_3d import FractureNetwork3d
 from porepy.grids.grid import Grid
 from porepy.grids.simplex import TriangleGrid
 from porepy.grids.structured import TensorGrid
+
+logger = logging.getLogger(__name__)
 
 
 def distort_grid_1d(
@@ -283,6 +286,236 @@ def remesh_1d(g_old: pp.Grid, num_nodes: int, tol: float = 1e-6) -> pp.Grid:
     g.update_boundary_node_tag()
 
     return g
+
+
+def mdg_refinement(
+    mdg: pp.MixedDimensionalGrid,
+    mdg_ref: pp.MixedDimensionalGrid,
+    tol: float = 1e-8,
+    mode: Literal["nested"] = "nested",
+) -> None:
+    """Wrapper for ``'coarse_fine_cell_mapping'`` in the grid data dictionaries
+    to construct mapping for grids in mixed-dimensional grid.
+
+    Adds a data dictionary entry to each grid in mdg.
+    The key is ``'coarse_fine_cell_mapping'``.
+
+    Currently, only nested refinement is supported; more general cases are also
+    possible.
+
+    ..rubric:: Acknowledgement
+
+        The code was contributed by Haakon Ervik.
+
+    Note:
+        No node prop is added to the reference grids in mdg_ref.
+
+    Parameters:
+        mdg: Coarse mixed-dimensional grid.
+        mdg_ref: Refined mixed-dimensional grid.
+        tol: ``default=1e-8``
+
+            Tolerance for point_in_poly* -methods
+        mode: ``default='nested'``
+
+            Refinement mode. ``'nested'`` corresponds to refinement by splitting.
+
+    """
+
+    # Check that both mdg's have the same dimension
+    dim_mdg = mdg.dim_max()
+    dim_mdg_ref = mdg_ref.dim_max()
+    assert dim_mdg_ref == dim_mdg
+
+    # Check that both mdg's have the same number of subdomains
+    subdomains = mdg.subdomains()
+    subdomains_ref = mdg_ref.subdomains()
+    assert len(subdomains) == len(subdomains_ref)
+
+    # Check that both mdg's have the same bounding box
+    bmin, bmax = pp.domain.mdg_minmax_coordinates(mdg)
+    bmin_ref, bmax_ref = pp.domain.mdg_minmax_coordinates(mdg_ref)
+    np.testing.assert_almost_equal(bmin, bmin_ref, decimal=8)
+    np.testing.assert_almost_equal(bmax, bmax_ref, decimal=8)
+
+    # Strongly check that grids refer to same domain
+    for i in np.arange(len(subdomains)):
+        sd, sd_ref = subdomains[i], subdomains_ref[i]
+        assert sd.id == sd_ref
+
+        # Compute the mapping for this subdomain-pair, and assign the result to the node
+        # of the coarse mdg
+        if mode == "nested":
+            mapping = structured_refinement(sd, sd_ref, point_in_poly_tol=tol)
+        else:
+            raise NotImplementedError("Unknown refinement mode")
+
+        mdg.subdomain_data(sd)["coarse_fine_cell_mapping"] = mapping
+
+
+def structured_refinement(
+    g: pp.Grid, g_ref: pp.Grid, point_in_poly_tol: float = 1e-8
+) -> sps.csc_matrix:
+    """Construct a mapping between cells of a grid and its refined version.
+
+    Assuming a regular and a refined mesh, where the refinement is executed by
+    splitting.
+    I.e. a cell in the refined grid is completely contained within a cell in the
+    coarse grid.
+
+    .. rubric:: Acknowledgement
+
+    The code was contributed by Haakon Ervik.
+
+    Parameters:
+        g: Coarse grid.
+        g_ref: Refined grid.
+        point_in_poly_tol: ``default=1e-8``
+
+            Tolerance for :func:`~porepy.geometry_property_checks.point_in_polyhedron`.
+
+    Raises:
+        AssertionError: If ``g_ref`` has more cells than ``g`` or if they are of unequal
+            dimension.
+        AssertionError: Depending on the dimension of ``g``, a certain number of nodes
+            must be available.
+        AssertionError: If any cell in the finer grid is not contained in exactly one
+            cell of the coarser grid.
+
+    Returns:
+        Column major sparse matrix mapping from coarse to fine cells.
+
+    """
+    # This method creates a mapping from fine to coarse cells by creating a matrix 'M'
+    # where the rows represent the fine cells while the columns represent the coarse
+    # cells. In practice this means that for an array 'p', of known values on a coarse
+    # grid, by applying the mapping
+    #     q = M * p
+    # each value in a coarse cell will now be transferred to all the fine cells
+    # contained within the coarse cell.
+
+    # The procedure for creating this mapping relies on two main assumptions.
+    #     1. Each fine cell is fully contained inside exactly one coarse cell.
+    #     2. Each cell can be characterised as a simplex.
+    #         I.e. Every node except one defines every face of the object.
+
+    # The first assumption implies that the problem of assessing if a fine cell is
+    # contained within a coarse cell is reduced to assessing if the center of a fine
+    # cell is contained within the coarse cell.
+
+    # The second assumption implies that a cell in any dimension (1D, 2D, 3D) will be
+    # a convex object. This way, we can use existing algorithms in PorePy to find if
+    # a point is inside a polygon (2D) or polyhedron (3D). (The 1D case is trivial)
+
+    # The general algorithm is as follows (refer to start of for-loop in the code):
+    # 1. Make a list of (pointers to) untested cell centers called 'test_cells_ptr'.
+    # 2. Iterate through all coarse cells. Now, consider one of these:
+    # 3. For all untested cell centers (defined by 'test_cells_ptr'), check if they
+    #     are inside the coarse cell.
+    # 4. Those that pass (is inside the coarse cell) add to the mapping, then remove
+    #     those point from the list of untested cell centers.
+    # 5. Assemble the mapping.
+
+    if g.dim == 0:
+        mapping = sps.csc_matrix((np.ones(1), ([0], [0])))
+        return mapping
+    assert g.num_cells < g_ref.num_cells, "Wrong order of input grids"
+    assert g.dim == g_ref.dim, "Grids must be of same dimension"
+
+    # 1. Step: Create a list of tuples pointing to the (start, end) index of the nodes
+    # of each cell on the coarse grid.
+    cell_nodes = g.cell_nodes()
+    # start/end row pointers for each column
+    nodes_of_cell_ptr = zip(cell_nodes.indptr[:-1], cell_nodes.indptr[1:])
+
+    # 2. Step: Initialize a sps.csc_matrix mapping fine cells to coarse cells.
+    indptr = np.array([0])
+    indices = np.empty(0)
+
+    cells_ref = g_ref.cell_centers.copy()  # Cell centers in fine grid
+    test_cells_ptr = np.arange(g_ref.num_cells)  # Pointer to cell centers
+    nodes = g.nodes.copy()
+
+    # 3. Step: If the grids are in 1D or 2D, we can simplify the calculation by rotating
+    # the coordinate system to local coordinates. For example, a 2D grid embedded in 3D
+    # would be "rotated" so that each coordinate is of the form (x, y, 0).
+    if g.dim == 1:
+        # Rotate coarse nodes and fine cell centers to align with the x-axis
+        tangent = pp.map_geometry.compute_tangent(nodes)
+        reference = np.array([1.0, 0.0, 0.0])
+        R = pp.map_geometry.project_line_matrix(nodes, tangent, reference=reference)
+        nodes = R.dot(nodes)[0, :]
+        cells_ref = R.dot(cells_ref)[0, :]
+
+    elif g.dim == 2:  # Pre-processing for efficiency
+        # Rotate coarse nodes and fine cell centers to the xy-plane.
+        R = pp.map_geometry.project_plane_matrix(nodes, check_planar=False)
+        nodes = np.dot(R, nodes)[:2, :]
+        cells_ref = np.dot(R, cells_ref)[:2, :]
+
+    # 4. Step: Loop through every coarse cell
+    for st, nd in nodes_of_cell_ptr:
+        nodes_idx = cell_nodes.indices[st:nd]
+        num_nodes = nodes_idx.size
+
+        # 5. Step: for the appropriate grid dimension, test all cell centers not already
+        # found to be inside some other coarse cell.
+        # 'in_poly' is a boolean array that is True for the points inside and False for
+        # the points not inside.
+        if g.dim == 1:
+            # In 1D, we use a numpy method to check which coarse cell the fine points
+            # are inside.
+            assert num_nodes == 2, "We assume a 'cell' in 1D is defined by two points."
+            line = np.sort(nodes[nodes_idx])
+            test_points = cells_ref[test_cells_ptr]
+            in_poly = np.searchsorted(line, test_points, side="left") == 1
+
+        elif g.dim == 2:
+            assert num_nodes == 3, "We assume simplexes in 2D (i.e. 3 nodes)"
+            polygon = nodes[:, nodes_idx]
+            test_points = cells_ref[:, test_cells_ptr]
+            in_poly = pp.geometry_property_checks.point_in_polygon(
+                poly=polygon, p=test_points
+            )
+
+        elif g.dim == 3:
+            # Make polyhedron from node coordinates.
+            # Polyhedron defined as a list of nodes defining its (convex) faces.
+            # Assumes simplexes: Every node except one defines every face.
+            assert num_nodes == 4, "We assume simplexes in 3D (i.e. 4 nodes)"
+            node_coords = nodes[:, nodes_idx]
+
+            ids = np.arange(num_nodes)
+            polyhedron: np.ndarray = np.array(
+                [node_coords[:, ids != i] for i in np.arange(num_nodes)]
+            )
+            # Test only points not inside another polyhedron.
+            test_points = cells_ref[:, test_cells_ptr]
+            in_poly = pp.geometry_property_checks.point_in_polyhedron(
+                polyhedron=polyhedron, test_points=test_points, tol=point_in_poly_tol
+            )
+
+        else:
+            logger.warning("A grid of dimension %d encountered. Skip!", g.dim)
+            continue
+
+        # 6. Step: Update pointer to which cell centers to use as test points
+        in_poly_ids = test_cells_ptr[in_poly]  # id of cells inside this polyhedron
+        # Keep only cells not inside this polyhedron
+        test_cells_ptr = test_cells_ptr[~in_poly]
+
+        # Update mapping
+        indices = np.append(indices, in_poly_ids)
+        indptr = np.append(indptr, indptr[-1] + in_poly_ids.size)
+
+    # 7. Step: assemble the sparse matrix with the mapping.
+    data = np.ones(indices.size)
+    coarse_fine = sps.csc_matrix((data, indices, indptr))
+
+    assert (
+        indices.size == g_ref.num_cells
+    ), "Every fine cell should be inside exactly one coarse cell"
+    return coarse_fine
 
 
 class GridSequenceIterator:
